+++ app/views/tweets/show.html.haml
@@ -33,3 +33,5 @@
   = link_to new_tweet_comment_path(@tweet) do
     .btn-secondary
       コメントを追加
+
+= javascript_pack_tag 'tweet'
